refactor: add admin service interface
