+++ AGENTS.md
@@ -27,6 +27,8 @@
 - `npm run migration:down` – revert the last migration
 - Do not commit `node_modules` or `package-lock.json`.
 - When environment variables change, update `.env.example` accordingly.
+- When adding new services, define an interface and export an Inversify key (Symbol)
+  so they can be bound in the container.
 
 ## Troubleshooting
